Conflicts:
	Makefile

+++ include/libopencm3/stm32/f1/memorymap.h
@@ -25,7 +25,7 @@
 /* --- STM32 specific peripheral definitions ------------------------------- */
 
 /* Memory map for all busses */
-#define PERIPH_BASE			0x40000000
+#define PERIPH_BASE			((u32)0x40000000)
 #define PERIPH_BASE_APB1		(PERIPH_BASE + 0x00000)
 #define PERIPH_BASE_APB2		(PERIPH_BASE + 0x10000)
 #define PERIPH_BASE_AHB			(PERIPH_BASE + 0x18000)
